backend/requirements.txt
fastapi==0.115.5
uvicorn==0.32.1
openai==1.55.3
pydantic==2.10.3
python-dotenv==1.0.1
PyJWT==2.9.0
cryptography==44.0.0
requests==2.32.3
